# agent-studio-server/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "agent swagger",
    "version" : "v1",
    "title" : "swagger"
  },
  "basePath" : "/",
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/api/v1/chat/completions" : {
      "post" : {
        "operationId" : "createChatCompletionStream",
        "consumes" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/ChatCompletionRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/FluxChatCompletionStreamResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
    "ChatCompletionRequest" : {
      "type" : "object",
      "properties" : {
        "model" : {
          "type" : "string"
        },
        "messages" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Message"
          }
        },
        "temperature" : {
          "type" : "number",
          "format" : "double"
        },
        "stream" : {
          "type" : "boolean"
        },
        "stop" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "user" : {
          "type" : "string"
        },
        "tools" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Tool"
          }
        },
        "seed" : {
          "type" : "integer",
          "format" : "int32"
        },
        "model_config_id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "top_p" : {
          "type" : "number",
          "format" : "double"
        },
        "max_tokens" : {
          "type" : "integer",
          "format" : "int32"
        },
        "presence_penalty" : {
          "type" : "number",
          "format" : "double"
        },
        "frequency_penalty" : {
          "type" : "number",
          "format" : "double"
        },
        "tool_choice" : {
          "type" : "object"
        },
        "response_format" : {
          "$ref" : "#/definitions/ResponseFormat"
        }
      }
    },
...
    "FluxChatCompletionStreamResponse" : {
      "type" : "object",
      "properties" : {
        "prefetch" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "Function" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "parameters" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        }
      }
    },
    "FunctionCall" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "arguments" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        }
      }
    },
    "Message" : {
      "type" : "object",
      "properties" : {
        "role" : {
          "type" : "string"
        },
        "content" : {
          "type" : "string"
        },
        "name" : {
          "type" : "string"
        },
        "tool_calls" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ToolCall"
          }
        },
        "tool_call_id" : {
          "type" : "string"
        }
      }
    },
...
    "ResponseFormat" : {
      "type" : "object",
      "properties" : {
        "type" : {
          "type" : "string"
        },
        "json_schema" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "object"
          }
        }
      }
    },
    "Tool" : {
      "type" : "object",
      "properties" : {
        "type" : {
          "type" : "string"
        },
        "function" : {
          "$ref" : "#/definitions/Function"
        }
      }
    },
    "ToolCall" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        },
        "function" : {
          "$ref" : "#/definitions/FunctionCall"
        }
      }
    }
  }
}